Trait srml_system::offchain::CreateTransaction

pub trait CreateTransaction<T: Trait, Extrinsic: ExtrinsicT> {
    type Signature;
    fn create_transaction<F: Signer<T::AccountId, Self::Signature>>(
        call: Extrinsic::Call, 
        account: T::AccountId, 
        nonce: T::Index
    ) -> Option<(Extrinsic::Call, Extrinsic::SignaturePayload)>;
}

Creates runtime-specific signed transaction.

Associated Types

type Signature

Loading content...
Required methods

fn create_transaction<F: Signer<T::AccountId, Self::Signature>>(
call: Extrinsic::Call,
account: T::AccountId,
nonce: T::Index
) -> Option<(Extrinsic::Call, Extrinsic::SignaturePayload)>

Attempt to create signed extrinsic data that encodes call from given account.

Runtime implementation is free to construct the payload to sign and the signature in any way it wants. Returns None if signed extrinsic could not be created (either because signing failed or because of any other runtime-specific reason).

Loading content...