Trait srml_system::offchain::Signer

pub trait Signer<Account, Signature> {
    fn sign<Payload: Encode>(
        account: Account, 
        payload: &Payload
    ) -> Option<Signature>;
}

A trait responsible for signing a payload using given account.

Required methods

fn sign<Payload: Encode>(
account: Account,
payload: &Payload
) -> Option<Signature>

Sign any encodable payload with given account and produce a signature.

Returns Some if signing succeeded and None in case the account couldn't be used.

Loading content...