Struct srml_system::Module

pub struct Module<T: Trait>(_);

Methods

impl<T: Trait> Module<T>

pub fn deposit_event(event: T::Event)

impl<T: Trait> Module<T>

Can also be called using Call.

pub fn set_code(origin: T::Origin, new: Vec<u8>) -> Result

Set the new code.

impl<T: 'static + Trait> Module<T>

pub fn account_nonce<K: Borrow<T::AccountId>>(key: K) -> T::Index

Extrinsics nonce for accounts.

pub fn next_weight_multiplier() -> WeightMultiplier

The next weight multiplier. This should be updated at the end of each block based on the saturation level (weight).

pub fn block_hash<K: Borrow<T::BlockNumber>>(key: K) -> T::Hash

Map of block numbers to block hashes.

pub fn extrinsic_data<K: Borrow<u32>>(key: K) -> Vec<u8>

Extrinsics data for the current block (maps an extrinsic's index to its data).

pub fn random_material() -> (i8, Vec<T::Hash>)

Series of block headers from the last 81 blocks that acts as random seed material. This is arranged as a ring buffer with the i8 prefix being the index into the Vec of the oldest hash.

pub fn block_number() -> T::BlockNumber

The current block number being processed. Set by execute_block.

pub fn parent_hash() -> T::Hash

Hash of the previous block.

pub fn extrinsics_root() -> T::Hash

Extrinsics root of the current block, also part of the block header.

pub fn digest() -> DigestOf<T>

Digest of the current block, also part of the block header.

pub fn events() -> Vec<EventRecord<T::Event, T::Hash>>

Events deposited for the current block.

pub fn event_count() -> u32

The number of events in the Events<T> list.

pub fn event_topics<KArg1: ?Sized, KArg2: ?Sized>(
k1: &KArg1,
k2: &KArg2
) -> Vec<(T::BlockNumber, u32)> where
(): Borrow<KArg1>,
T::Hash: Borrow<KArg2>,
KArg1: Encode,
KArg2: Encode,

impl<T: Trait> Module<T>

pub fn deposit_event_indexed(topics: &[T::Hash], event: T::Event)

Deposits an event into this block's event record adding this event to the corresponding topic indexes.

This will update storage entries that correspond to the specified topics. It is expected that light-clients could subscribe to this topics.

pub fn extrinsic_index() -> Option<u32>

Gets the index of extrinsic that is currently executing.

pub fn extrinsic_count() -> u32

Gets extrinsics count.

pub fn all_extrinsics_weight() -> Weight

Gets a total weight of all executed extrinsics.

pub fn all_extrinsics_len() -> u32

pub fn update_weight_multiplier()

Update the next weight multiplier.

This should be called at then end of each block, before all_extrinsics_weight is cleared.

pub fn initialize(
number: &T::BlockNumber,
parent_hash: &T::Hash,
txs_root: &T::Hash,
digest: &DigestOf<T>
)

Start the execution of a particular block.

pub fn finalize() -> T::Header

Remove temporary "environment" entries in storage.

pub fn deposit_log(item: DigestItemOf<T>)

Deposits a log and ensures it matches the block's log data.

pub fn externalities() -> TestExternalities<Blake2Hasher>

Get the basic externalities for this module, useful for tests.

pub fn set_block_number(n: T::BlockNumber)

Set the block number to something in particular. Can be used as an alternative to initialize for tests that don't need to bother with the other environment entries.

pub fn set_extrinsic_index(extrinsic_index: u32)

Sets the index of extrinsic that is currently executing.

pub fn set_parent_hash(n: T::Hash)

Set the parent hash number to something in particular. Can be used as an alternative to initialize for tests that don't need to bother with the other environment entries.

pub fn random_seed() -> T::Hash

Get the basic random seed.

In general you won't want to use this, but rather Self::random which allows you to give a subject for the random result and whose value will be independently low-influence random from any other such seeds.

pub fn random(subject: &[u8]) -> T::Hash

Get a low-influence "random" value.

Being a deterministic block chain, real randomness is difficult to come by. This gives you something that approximates it. subject is a context identifier and allows you to get a different result to other callers of this function; use it like random(&b"my context"[..]).

This is initially implemented through a low-influence "triplet mix" convolution of previous block hash values. In the future it will be generated from a secure verifiable random function (VRF).

Security Notes

This randomness uses a low-influence function, drawing upon the block hashes from the previous 81 blocks. Its result for any given subject will be known in advance by the block producer of this block (and, indeed, anyone who knows the block's parent_hash). However, it is mostly impossible for the producer of this block alone to influence the value of this hash. A sizable minority of dishonest and coordinating block producers would be required in order to affect this value. If that is an insufficient security guarantee then two things can be used to improve this randomness:

Name, in advance, the block number whose random value will be used; ensure your module retains a buffer of previous random values for its subject and then index into these in order to obviate the ability of your user to look up the parent hash and choose when to transact based upon it.

Require your user to first commit to an additional value by first posting its hash. Require them to reveal the value to determine the final result, hashing it with the output of this random function. This reduces the ability of a cabal of block producers from conspiring against individuals.

WARNING: Hashing the result of this function will remove any low-influnce properties it has and mean that all bits of the resulting value are entirely manipulatable by the author of the parent block, who can determine the value of parent_hash.

pub fn inc_account_nonce(who: &T::AccountId)

Increment a particular account's nonce by 1.

pub fn note_extrinsic(encoded_xt: Vec<u8>)

Note what the extrinsic data of the current extrinsic index is. If this is called, then ensure derive_extrinsics is also called before block-building is completed.

NOTE: This function is called only when the block is being constructed locally. execute_block doesn't note any extrinsics.

pub fn note_applied_extrinsic(r: &Result<(), &'static str>, _encoded_len: u32)

To be called immediately after an extrinsic has been applied.

pub fn note_finished_extrinsics()

To be called immediately after note_applied_extrinsic of the last extrinsic of the block has been called.

pub fn derive_extrinsics()

Remove all extrinsic data and save the extrinsics trie root.