Trait srml_system::Trait

pub trait Trait: 'static + Eq + Clone {
    type Origin: Into<Result<RawOrigin<Self::AccountId>, Self::Origin>> + From<RawOrigin<Self::AccountId>>;
    type Index: Parameter + Member + MaybeSerializeDebugButNotDeserialize + Default + MaybeDisplay + SimpleArithmetic + Copy;
    type BlockNumber: Parameter + Member + MaybeSerializeDebug + MaybeDisplay + SimpleArithmetic + Default + Bounded + Copy + Hash;
    type Hash: Parameter + Member + MaybeSerializeDebug + MaybeDisplay + SimpleBitOps + Default + Copy + CheckEqual + Hash + AsRef<[u8]> + AsMut<[u8]>;
    type Hashing: Hash<Output = Self::Hash>;
    type AccountId: Parameter + Member + MaybeSerializeDebug + MaybeDisplay + Ord + Default;
    type Lookup: StaticLookup<Target = Self::AccountId>;
    type Header: Parameter + Header<Number = Self::BlockNumber, Hash = Self::Hash>;
    type Event: Parameter + Member + From<Event>;
}

Associated Types

type Origin: Into<Result<RawOrigin<Self::AccountId>, Self::Origin>> + From<RawOrigin<Self::AccountId>>

The aggregated Origin type used by dispatchable calls.

type Index: Parameter + Member + MaybeSerializeDebugButNotDeserialize + Default + MaybeDisplay + SimpleArithmetic + Copy

Account index (aka nonce) type. This stores the number of previous transactions associated with a sender account.

type BlockNumber: Parameter + Member + MaybeSerializeDebug + MaybeDisplay + SimpleArithmetic + Default + Bounded + Copy + Hash

The block number type used by the runtime.

type Hash: Parameter + Member + MaybeSerializeDebug + MaybeDisplay + SimpleBitOps + Default + Copy + CheckEqual + Hash + AsRef<[u8]> + AsMut<[u8]>

The output of the Hashing function.

type Hashing: Hash<Output = Self::Hash>

The hashing system (algorithm) being used in the runtime (e.g. Blake2).

type AccountId: Parameter + Member + MaybeSerializeDebug + MaybeDisplay + Ord + Default

The user account identifier type for the runtime.

type Lookup: StaticLookup<Target = Self::AccountId>

Converting trait to take a source type and convert to AccountId.

Used to define the type and conversion mechanism for referencing accounts in transactions. It's perfectly reasonable for this to be an identity conversion (with the source type being AccountId), but other modules (e.g. Indices module) may provide more functional/efficient alternatives.

type Header: Parameter + Header<Number = Self::BlockNumber, Hash = Self::Hash>

The block header.

type Event: Parameter + Member + From<Event>

The aggregated event type of the runtime.

Loading content...